JDK 1.5 update, uprava cesty v WhatHappened
git-svn-id: https://zdrojaky.abclinuxu.cz/svn-abclinuxu/trunk@1690 be82d567-0701-4079-a524-1ea639e27e29

+++ src/cz/abclinuxu/utils/feeds/FeedGenerator.java
@@ -238,7 +238,7 @@ public class FeedGenerator implements Configurable {
             SyndContent description;
 
             Map defaultSizes = VariableFetcher.getInstance().getDefaultSizes();
-            int countArticles = ((Integer)defaultSizes.get(VariableFetcher.KEY_ARTICLE)).intValue();
+            int countArticles = (Integer) defaultSizes.get(VariableFetcher.KEY_ARTICLE);
             Qualifier[] qualifiers = new Qualifier[]{Qualifier.SORT_BY_CREATED, Qualifier.ORDER_DESCENDING, new LimitQualifier(0, countArticles)};
             List articles = sqlTool.findIndexArticlesRelations(qualifiers);
             Tools.syncList(articles);
